refactor(globals): extract global types and constants
