+++ firebase/utilityFunctions.ts
@@ -1,4 +1,4 @@
-import { TNullableNumber } from './documentsDataTypes'
+import { TNullableNumber } from '../globals/types'
 
 const LEN_SIMPLE_COURSE_NUMBER = 5 //   DD-CCCC     (e.g., CS-6200)
 const LEN_COMPOUND_COURSE_NUMBER = 6 // DD-CCCC-CCC (e.g., CS-8803-O08)
