+++ globals/constants.ts
@@ -1,0 +1,10 @@
+import { TKeyMap } from './types'
+
+export const idKeys: TKeyMap = {
+	COURSE_ID: 'courseId',
+	DEPARTMENT_ID: 'departmentId',
+	PROGRAM_ID: 'programId',
+	SEMESTER_ID: 'semesterId',
+	SPECIALIZATION_ID: 'specializationId',
+	USER_ID: 'userId',
+}
